ppg-8 (TDD) ChangerOrdreEtapeService : passage en type long wrapper EtapeWorkflowDTO : ajout d'un attribut id OrdreEtapeDTO : changement type int en long ProjetRepository : bugs fixe (suppression espace et ; dans la query) UcmppgApplication : ajout du droit ChangerOrdreEtape dans les droits de bases d'un administrateur

+++ src/main/java/be/technobel/ucmppg/bl/dto/projet/workflow/OrdreEtapeDTO.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class OrdreEtapeDTO {
 
-    private int idUtilisateur;
+    private long idUtilisateur;
     private int nvOrdre;
 }
